Move `CreateActionForm` to `ActionCreator`
